// frontend/src/lib/api/generated/models/issueDetailResponse.ts
/**
 * Generated by Orval. Do not edit manually.
 */
import type { Issue } from "./issue.ts";
import type { IssueEvent } from "./issueEvent.ts";
import type { RepoRefResponse } from "./repoRefResponse.ts";
import type { WorkflowStateMetaResponse } from "./workflowStateMetaResponse.ts";
import type { WorkspaceRef } from "./workspaceRef.ts";

export interface IssueDetailResponse {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  detail_fetched_at?: string;
  detail_loaded: boolean;
  events: IssueEvent[];
  issue: Issue;
  platform_host: string;
  repo: RepoRefResponse;
  repo_name: string;
  repo_owner: string;
  workflow?: WorkflowStateMetaResponse;
  workspace?: WorkspaceRef;
}